feat: melhorando as validações do Back-end

+++ src/app/controllers/CategoryController.js
@@ -1,4 +1,5 @@
 const CategoriesRepository = require('../repositories/CategoriesRepository');
+const isValidUUID = require('../utils/isValidUUID');
 
 class CategoryController {
   async index(request, response) {
@@ -8,6 +9,11 @@ class CategoryController {
 
   async show(request, response) {
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid category id' });
+    }
+
     const category = await CategoriesRepository.findById(id);
     response.status(200).json(category);
   }
@@ -27,13 +33,18 @@ class CategoryController {
   async update(request, response) {
     const { id } = request.params;
     const { name } = request.body;
-    const categoryExists = await CategoriesRepository.findById(id);
-    if (!categoryExists) {
-      return response.status(404).json({ error: 'Category not found' });
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid category id' });
     }
 
     if (!name) {
       return response.status(400).json({ error: 'Name is required' });
+    }
+
+    const categoryExists = await CategoriesRepository.findById(id);
+    if (!categoryExists) {
+      return response.status(404).json({ error: 'Category not found' });
     }
 
     const updatedCategory = await CategoriesRepository.update(id, { name });
@@ -42,6 +53,11 @@ class CategoryController {
 
   async delete(request, response) {
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid category id' });
+    }
+
     await CategoriesRepository.delete(id);
     response.sendStatus(204);
   }
